+++ admin/src/main/java/cz/quantumleap/admin/AdminController.java
@@ -5,8 +5,8 @@ import cz.quantumleap.admin.menu.AdminMenuItem.State;
 import cz.quantumleap.admin.menu.AdminMenuManager;
 import cz.quantumleap.admin.notification.NotificationService;
 import cz.quantumleap.admin.person.PersonService;
-import cz.quantumleap.core.notification.transport.Notification;
-import cz.quantumleap.core.person.transport.Person;
+import cz.quantumleap.core.notification.domain.Notification;
+import cz.quantumleap.core.person.domain.Person;
 import cz.quantumleap.core.security.WebSecurityExpressionEvaluator;
 import org.springframework.security.core.Authentication;
 import org.springframework.util.AntPathMatcher;
