+++ admin/src/main/java/cz/quantumleap/admin/notification/NotificationController.java
@@ -4,10 +4,10 @@ import cz.quantumleap.admin.AdminController;
 import cz.quantumleap.admin.menu.AdminMenuManager;
 import cz.quantumleap.admin.person.PersonService;
 import cz.quantumleap.core.common.Utils;
-import cz.quantumleap.core.database.transport.SliceRequest;
-import cz.quantumleap.core.database.transport.TableSlice;
-import cz.quantumleap.core.notification.transport.Notification;
-import cz.quantumleap.core.person.transport.Person;
+import cz.quantumleap.core.database.domain.SliceRequest;
+import cz.quantumleap.core.database.domain.TableSlice;
+import cz.quantumleap.core.notification.domain.Notification;
+import cz.quantumleap.core.person.domain.Person;
 import cz.quantumleap.core.security.WebSecurityExpressionEvaluator;
 import cz.quantumleap.core.view.DefaultListController;
 import org.springframework.security.core.Authentication;
