+++ admin/src/main/java/cz/quantumleap/admin/person/PersonController.java
@@ -8,13 +8,13 @@ import cz.quantumleap.admin.notification.NotificationService;
 import cz.quantumleap.admin.personrole.PersonRoleController;
 import cz.quantumleap.admin.personrole.PersonRoleService;
 import cz.quantumleap.core.common.Utils;
+import cz.quantumleap.core.database.domain.SliceRequest;
+import cz.quantumleap.core.database.domain.TableSlice;
 import cz.quantumleap.core.database.entity.EntityIdentifier;
-import cz.quantumleap.core.database.transport.SliceRequest;
-import cz.quantumleap.core.database.transport.TableSlice;
-import cz.quantumleap.core.person.transport.Person;
+import cz.quantumleap.core.person.domain.Person;
 import cz.quantumleap.core.security.WebSecurityExpressionEvaluator;
 import cz.quantumleap.core.session.SessionService;
-import cz.quantumleap.core.session.transport.SessionDetail;
+import cz.quantumleap.core.session.domain.SessionDetail;
 import cz.quantumleap.core.tables.PersonRoleTable;
 import cz.quantumleap.core.view.*;
 import org.springframework.beans.factory.annotation.Qualifier;
