+++ admin/src/main/java/cz/quantumleap/admin/personrole/PersonRoleService.java
@@ -2,7 +2,7 @@ package cz.quantumleap.admin.personrole;
 
 import cz.quantumleap.core.business.ServiceStub;
 import cz.quantumleap.core.personrole.PersonRoleDao;
-import cz.quantumleap.core.personrole.transport.PersonRole;
+import cz.quantumleap.core.personrole.domain.PersonRole;
 import org.springframework.stereotype.Service;
 
 @Service
