+++ admin/src/main/java/cz/quantumleap/admin/role/RoleController.java
@@ -6,8 +6,8 @@ import cz.quantumleap.admin.menu.AdminMenuItemDefinition;
 import cz.quantumleap.admin.menu.AdminMenuManager;
 import cz.quantumleap.admin.notification.NotificationService;
 import cz.quantumleap.admin.person.PersonService;
-import cz.quantumleap.core.database.transport.SliceRequest;
-import cz.quantumleap.core.role.transport.Role;
+import cz.quantumleap.core.database.domain.SliceRequest;
+import cz.quantumleap.core.role.domain.Role;
 import cz.quantumleap.core.security.WebSecurityExpressionEvaluator;
 import cz.quantumleap.core.view.*;
 import org.springframework.security.access.prepost.PreAuthorize;
