+++ admin/src/main/java/cz/quantumleap/admin/web/TemplateHelper.java
@@ -1,6 +1,6 @@
 package cz.quantumleap.admin.web;
 
-import cz.quantumleap.core.database.transport.TableSlice;
+import cz.quantumleap.core.database.domain.TableSlice;
 import cz.quantumleap.core.view.controllerargument.SliceRequestControllerArgumentResolver;
 
 public class TemplateHelper {
