+++ core/src/main/java/cz/quantumleap/core/autoincrement/domain/Increment.java
@@ -1,4 +1,4 @@
-package cz.quantumleap.core.autoincrement.transport;
+package cz.quantumleap.core.autoincrement.domain;
 
 public class Increment {
