+++ core/src/main/java/cz/quantumleap/core/business/LookupListServiceStub.java
@@ -2,9 +2,9 @@ package cz.quantumleap.core.business;
 
 import cz.quantumleap.core.database.ListDao;
 import cz.quantumleap.core.database.LookupDao;
+import cz.quantumleap.core.database.domain.SliceRequest;
+import cz.quantumleap.core.database.domain.TableSlice;
 import cz.quantumleap.core.database.entity.EntityIdentifier;
-import cz.quantumleap.core.database.transport.SliceRequest;
-import cz.quantumleap.core.database.transport.TableSlice;
 import org.jooq.Record;
 import org.jooq.Table;
