+++ core/src/main/java/cz/quantumleap/core/database/LookupListDaoStub.java
@@ -1,8 +1,8 @@
 package cz.quantumleap.core.database;
 
+import cz.quantumleap.core.database.domain.SliceRequest;
+import cz.quantumleap.core.database.domain.TableSlice;
 import cz.quantumleap.core.database.entity.Entity;
-import cz.quantumleap.core.database.transport.SliceRequest;
-import cz.quantumleap.core.database.transport.TableSlice;
 import org.jooq.Condition;
 import org.jooq.DSLContext;
 import org.jooq.Record;
