+++ core/src/main/java/cz/quantumleap/core/database/domain/IdLabel.java
@@ -1,4 +1,4 @@
-package cz.quantumleap.core.database.transport;
+package cz.quantumleap.core.database.domain;
 
 public class IdLabel {
