+++ core/src/main/java/cz/quantumleap/core/database/query/LimitFactory.java
@@ -1,6 +1,6 @@
 package cz.quantumleap.core.database.query;
 
-import cz.quantumleap.core.database.transport.SliceRequest;
+import cz.quantumleap.core.database.domain.SliceRequest;
 
 public final class LimitFactory {
