+++ core/src/main/java/cz/quantumleap/core/notification/NotificationDao.java
@@ -3,10 +3,10 @@ package cz.quantumleap.core.notification;
 import cz.quantumleap.core.database.DaoStub;
 import cz.quantumleap.core.database.EntityRegistry;
 import cz.quantumleap.core.database.RecordAuditor;
+import cz.quantumleap.core.database.domain.SliceRequest;
+import cz.quantumleap.core.database.domain.TableSlice;
 import cz.quantumleap.core.database.entity.Entity;
-import cz.quantumleap.core.database.transport.SliceRequest;
-import cz.quantumleap.core.database.transport.TableSlice;
-import cz.quantumleap.core.notification.transport.Notification;
+import cz.quantumleap.core.notification.domain.Notification;
 import cz.quantumleap.core.tables.NotificationTable;
 import org.jooq.Condition;
 import org.jooq.DSLContext;
