+++ core/src/main/java/cz/quantumleap/core/notification/NotificationManager.java
@@ -1,6 +1,6 @@
 package cz.quantumleap.core.notification;
 
-import cz.quantumleap.core.notification.transport.Notification;
+import cz.quantumleap.core.notification.domain.Notification;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
