+++ core/src/main/java/cz/quantumleap/core/personrole/domain/PersonRole.java
@@ -1,4 +1,4 @@
-package cz.quantumleap.core.personrole.transport;
+package cz.quantumleap.core.personrole.domain;
 
 import javax.validation.constraints.NotNull;
