+++ core/src/main/java/cz/quantumleap/core/role/domain/Role.java
@@ -1,4 +1,4 @@
-package cz.quantumleap.core.role.transport;
+package cz.quantumleap.core.role.domain;
 
 public class Role {
