+++ core/src/main/java/cz/quantumleap/core/view/ListController.java
@@ -1,6 +1,6 @@
 package cz.quantumleap.core.view;
 
-import cz.quantumleap.core.database.transport.SliceRequest;
+import cz.quantumleap.core.database.domain.SliceRequest;
 import org.springframework.ui.Model;
 
 import javax.servlet.http.HttpServletRequest;
